+++ src/main/java/com/digitalkrapht/bloodbank/bloodbank/users/request/BloodRecipientRequest.java
@@ -18,7 +18,7 @@ public class BloodRecipientRequest {
     @NotNull
     private String idNumber;
     @NotNull
-    private long bloodGroupId;
+    private int bloodGroupId;
     @NotNull
     private Gender gender;
